Implement Curve<T>.GetSlice

+++ client/example/Example.cs
@@ -5,7 +5,7 @@ public partial class Example : Node2D
 {
 	public Example() {
 		GD.Print("Hello World");
-		Downflux.Lib.Curve<int> x = new();
+		Downflux.Lib.Curve<float> x = new();
 		GD.Print(x.Get(Time.GetTicksMsec()));
 	}
